==> preprocessing_rnapsec/4_rnap_prepro_is.py <==

import pyEQL
import pandas as pd
import numpy as np
import math



def calc_salt2(df, df_salt, salt_conc2, salt_unit2):
    #saltの種類が一種類のやつからやってみる
    df_salt1 = df_salt[~df_salt.salt_conc.str.contains(",", na = True)]
    salt1_unit = df_salt1.salt_unit.str.split("/", expand = True)
    salt1_unit = salt1_unit.rename(columns = {0 : "unit", 1: "name"})
    df_salt1_1 = pd.concat([df_salt1, salt1_unit], axis = "columns")

    #pyEQLのpH：numじゃないと無理
    salt_ionic_strength_ion=[]
    salt1_index=[]
    for i in df_salt1_1.index:
        a=df_salt1_1.salt_conc[i]+" "+df_salt1_1.unit[i]
        b = df_salt1_1.name[i]
        
        e=(float(df_salt1_1.salt_conc[i])*2) #2価なので、濃度2倍に設定
        f=str(e)+" "+df_salt1_1.unit[i]
        if b=="NaCl" :
            #pHなし
            if pd.isnull(df_salt1_1.pH.loc[i]):
            #df_salt1_1.pH.loc[i].isna():
                salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]])
                salt_ionic_strength_ion.append(salt_a.get_ionic_strength())
                salt1_index.append(i)
            #pHあり
            else:
                c=math.floor(df_salt1_1.pH.loc[i]) 
                #print(i)
                salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]],pH=c)
                salt_ionic_strength_ion.append(salt_a.get_ionic_strength())
                salt1_index.append(i)
        elif b=="KCl":
            #pHなし
            if pd.isnull(df_salt1_1.pH.loc[i]):
            #df_salt1_1.pH.loc[i].isna():
                salt_a=pyEQL.Solution([["K+",a],["Cl-",a]])
                salt_ionic_strength_ion.append(salt_a.get_ionic_strength())
                salt1_index.append(i)
            #pHあり
            else:
                c=math.floor(df_salt1_1.pH.loc[i]) 
                #print(i)
                salt_a=pyEQL.Solution([["K+",a],["Cl-",a]], pH=c)
                salt_ionic_strength_ion.append(salt_a.get_ionic_strength())
                salt1_index.append(i)
                
        elif b == "Zn2+":
            c=math.floor(df_salt1_1.pH.loc[i]) 
            salt_a=pyEQL.Solution([["Zn2+",a],],pH=c)
            salt_ionic_strength_ion.append(salt_a.get_ionic_strength())
            salt1_index.append(i)
            
        #塩が計算できないやつ（尿素？とか）
        else:
            salt_ionic_strength_ion.append(None)
            salt1_index.append(i)

    salt1_dict={}
    for i in range(len(salt_ionic_strength_ion)):
        if salt_ionic_strength_ion[i]!=None:
            #print(salt1_index[i])
            a=str(salt_ionic_strength_ion[i].magnitude)
            salt1_dict[salt1_index[i]]=a
        else:
            salt1_dict[salt1_index[i]]=None
    df_a=pd.DataFrame(salt1_dict,index=salt1_dict.keys())
    df_salt1_is=pd.DataFrame(df_a.T.iloc[:,0])
    df_salt1_is=df_salt1_is.rename(columns={df_salt1_is.columns[0]:0})

    ##2種塩
    unit1 = salt_unit2[0].str.split("/", expand = True)[0]
    name1 = salt_unit2[0].str.split("/", expand = True)[1]
    unit2 = salt_unit2[1].str.split("/", expand = True)[0]
    name2 = salt_unit2[1].str.split("/", expand = True)[1]
    conc1 = salt_conc2[0]
    conc2 = salt_conc2[1]

    salt2_is = []
    salt2_index = []
    for i in salt_unit2.index:
        #2価のイオンはname2の方にしかないです
        a=conc1.loc[i]+" "+unit1.loc[i]
        d=conc2.loc[i]+" "+unit2.loc[i] #2価のイオンを含む
        e=(float(conc2.loc[i])*2) #2価なので、濃度2倍に設定
        f=str(e)+" "+unit2.loc[i]


        if name1.loc[i]=="NaCl" :
            
            if name2.loc[i] == "MgCl2" :
                #pHなし
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                    print(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            elif name2.loc[i] == 'ZnCl2':
                #pHなし
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Zn2+", d], ["Cl-",f]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Zn2+", d], ["Cl-",f]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            elif name2.loc[i] == 'KCl':
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["K+", d], ["Cl-",d]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Na+",d],["Cl-",d]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                    
            elif name2.loc[i] == 'NaCl':
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            else:
                salt2_is.append(salt_a.get_ionic_strength())
                salt2_index.append(i)
                
        elif name1.loc[i]=="KCl":
            if name2.loc[i] == "MgCl2" :
                #pHなし
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                    print(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            elif name2.loc[i] == 'ZnCl2':
                #pHなし
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Zn2+", d], ["Cl-",f]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Zn2+", d], ["Cl-",f]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            elif name2.loc[i] == 'KCl':
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["K+", d], ["Cl-",d]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Na+",d],["Cl-",d]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
                    
            elif name2.loc[i] == 'NaCl':
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]])
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a]],pH=c)
                    salt2_is.append(salt_a.get_ionic_strength())
                    salt2_index.append(i)
            
        #塩が計算できないやつ（尿素？とか）
        else:
            salt2_is.append(salt_a.get_ionic_strength())
            salt2_index.append(i)

    #リスト→辞書→pandas
    salt2_dict={}
    for i in range(len(salt_unit2)):

        aa=str(salt2_is[i].magnitude)
        salt2_dict[salt2_index[i]]=aa
    df_b=pd.DataFrame(salt2_dict,index=salt2_dict.keys())
    df_salt2_is=pd.DataFrame(df_b.T.iloc[:,0])
    df_salt2_is=df_salt2_is.rename(columns={df_salt2_is.columns[0]:0})
    
    return df_salt1_is, df_salt2_is

def calc_salt3(df, salt_conc, salt_unit):
    "3種塩"
    salt_conc3 = salt_conc[salt_conc[3].isna() & salt_conc[2].notna()]
    salt_unit3 = salt_unit[salt_unit.index.isin(salt_conc3.index)]
    salt_unit3[2] = salt_unit3[2].fillna("mM/ZnCl2") #(0704)欠損していたので、元ファイル参照に埋めた。(0708: 元ファイルの方を直したので、次多分消してもいい)
    #塩の種類ごとのconc
    conc3_0, conc3_1, conc3_2 = salt_conc3[0], salt_conc3[1], salt_conc3[2]
    #塩の種類ごとの単位
    unit3_0, unit3_1, unit3_2 = salt_unit3[0].str.split("/", expand = True)[0], salt_unit3[1].str.split("/", expand = True)[0], salt_unit3[2].str.split("/", expand = True)[0]
    #塩の種類
    name3_0, name3_1, name3_2 = salt_unit3[0].str.split("/", expand = True)[1], salt_unit3[1].str.split("/", expand = True)[1], salt_unit3[2].str.split("/", expand = True)[1]

    "計算"
    salt3_is = []
    salt3_index = []
    for i in salt_unit3.index:
        #2価のイオンはname3_1の方にしかないです
        a=conc3_0.loc[i]+" "+unit3_0.loc[i]
        d=conc3_2.loc[i]+" "+unit3_1.loc[i] #2価のイオンを含む
        e=(float(conc3_2.loc[i])*2) #2価なので、濃度2倍に設定
        f=str(e)+" "+unit3_1.loc[i]
        
        g = conc3_2.loc[i]+" "+unit3_2.loc[i]
        
        h2 = (float(conc3_2.loc[i])*2)
        h=str(h2)+" "+unit3_2.loc[i]
        #h: 3つ目の塩が2価の塩
        assert name3_0[(name3_0.str.contains("hospha"))| (name3_0.str.contains("NaCl"))].shape[0]== name3_0.shape[0], f"other salt included in name3_0: {name3_0.unique()}" # AttributeError: 'str' object has no attribute 'contains'
        salts = []
        if name3_0.loc[i]=="NaCl" :
            
            if (name3_1.loc[i] == "MgCl2") & (name3_2.loc[i] == "ZnCl2"):
                #pHなし MgCl2
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f], ["Zn2+", g], ["Cl", h]]) #g: 濃度1倍, h: 濃度2倍
                    salt3_is.append(salt_a.get_ionic_strength())
                    salt3_index.append(i)
                    print(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a],["Mg2+", d],["Cl-", f], ["Zn2+", g], ["Cl", h]], pH= c) #g: 濃度1倍, h: 濃度2倍
                    salt3_is.append(salt_a.get_ionic_strength())
                    salt3_index.append(i)
            if (name3_1.loc[i] == "KCl") & (name3_2.loc[i] == "MgCl2"):
                #pHなし
                if pd.isnull(df.pH.loc[i]):
                #ph3.pH.loc[f2["index"].loc[i]].isna():
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a], ["K+", d], ["Cl-",d ],["Mg2+", g],["Cl-", h]]) #g: 濃度1倍, h: 濃度2倍
                    salt3_is.append(salt_a.get_ionic_strength())
                    salt3_index.append(i)
                    print(i)
                #pHあり
                else:
                    c=math.floor(df.pH.loc[i]) 
                    #print(i)
                    salt_a=pyEQL.Solution([["Na+",a],["Cl-",a], ["K+", d], ["Cl-", d],["Mg2+", g],["Cl-", h]]) #g: 濃度1倍, h: 濃度2倍
                    salt3_is.append(salt_a.get_ionic_strength())
                    salt3_index.append(i)
            else:
                print("other salt?")
                print(0, name3_0.loc[i])
                print(1, name3_1.loc[i])
                print(2, name3_2.loc[i])
        #塩が計算できないやつ（尿素？とか）
        else:
            salt3_is.append(salt_a.get_ionic_strength())
            salt3_index.append(i)
    salt3_dict={}
    for i in range(len(salt3_is)):
        aa=str(salt3_is[i].magnitude)
        salt3_dict[salt3_index[i]]=aa

    df_b=pd.DataFrame(salt3_dict,index=salt3_dict.keys())
    df_salt3_is=pd.DataFrame(df_b.T.iloc[:,0])
    df_salt3_is=df_salt3_is.rename(columns={df_salt3_is.columns[0]:0})

    return df_salt3_is

def main():
    df = pd.read_csv("./preprocessing_results/rnap_preprocessing_1.csv", index_col=False)
    df = df.rename(columns = {"salt_unit/salt_name": "salt_unit_name"})
    df.pH[df.pH =="-"] = 7.4
    df.pH = df.pH.astype("float")
    df.salt_conc = df.salt_conc.replace({"-":np.nan}, )
    df.salt_conc = df.salt_conc.replace({None:np.nan})
    df.salt_unit_name = df.salt_unit_name.str.replace("ﾎｼ|μ", "u")
    df.salt_unit_name[df.salt_unit_name.str.contains("http", na = False)] = "mM/KCl, mM/MgCl2"
    df.salt_unit_name = df.salt_unit_name.replace({None:np.nan})
    df_salt = pd.DataFrame(df.salt_conc)
    df_salt["salt_unit"] = df.salt_unit_name
    df_salt["pH"] = df.pH.astype("float")

    salt_conc = df_salt.salt_conc.str.split(",", expand = True)
    salt_unit = df_salt.salt_unit.str.split(",", expand = True)

    salt_conc2 = salt_conc[salt_conc[2].isna() & salt_conc[1].notna()]
    salt_unit2 = salt_unit[salt_unit.index.isin(salt_conc2.index)]

    if df_salt.salt_unit.str.count(",").unique().shape[0] == 5:
        df_salt1_is, df_salt2_is = calc_salt2(df, df_salt, salt_conc2, salt_unit2)
        df_salt3_is = calc_salt3(df, salt_conc, salt_unit)
        is_list = [df_salt1_is, df_salt2_is, df_salt3_is]
    elif df_salt.salt_unit.str.count(",").unique().shape[0] == 4:
        df_salt1_is, df_salt2_is = calc_salt2(df, df_salt, salt_conc2, salt_unit2)
        is_list = [df_salt1_is, df_salt2_is,]
    "計算したイオン強度を大元のデータフレームにカラム追加する"

    df_salt_is = pd.concat(is_list, axis = "index")
    df["ionic_strength"] = df_salt_is

    df.ionic_strength = df.ionic_strength.fillna(0)
    df.ionic_strength[df.salt_unit_name.str.contains("phosphate", na = False)] = np.nan
    df.to_csv("./preprocessing_results/rnap_preprocessing_is.csv", index = False)

if __name__=="__main__":
    main()
